RavenDB-24909 feat(certificates renewal guidance modal): Modularize guidance content by splitting browser and API access instructions into separate components.
