+++ src/Raven.Studio/typescript/components/pages/resources/manageServer/certificates/partials/authEnabled/CertificatesRenewedModal.scss
@@ -2,72 +2,73 @@
 @use "Content/scss/colors";
 
 .certificates-renewed-modal {
-    // Top-level Browser Access / Application Access segmented tabs
-    .access-tab-container {
-        .nav {
-            gap: sizes.$gutter-sm;
-        }
-
-        .nav-link {
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            width: 100%;
-            color: colors.$text-emphasis-color;
-            font-size: sizes.$font-size-base;
-            text-align: center;
-            background-image: none !important;
-            border: 1px solid colors.$border-color-light-var;
-            border-radius: sizes.$border-radius;
-            padding: sizes.$gutter-xs sizes.$gutter-sm !important;
-            margin-bottom: 0 !important;
-            &[aria-selected="false"] {
-                color: colors.$text-muted-color;
-                background-color: colors.$body-bg-var;
-            }
-
-            &[aria-selected="true"] {
-                border-color: colors.$primary-var;
-                background-color: var(--faded-primary);
-            }
-        }
+  // Top-level Browser Access / Application Access segmented tabs
+  .access-tab-container {
+    .nav {
+      gap: sizes.$gutter-sm;
     }
 
-    // Segmented tab set (browser + client code) - mirrors .browser-tab-container
-    .segmented-tab-container {
-        border-radius: var(--bs-border-radius);
-        border: 1px solid var(--border-color-light);
-        background-color: var(--panel-bg-2);
+    .nav-link {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      width: 100%;
+      color: colors.$text-emphasis-color;
+      font-size: sizes.$font-size-base;
+      text-align: center;
+      background-image: none !important;
+      border: 1px solid colors.$border-color-light-var;
+      border-radius: sizes.$border-radius;
+      padding: sizes.$gutter-xs sizes.$gutter-sm !important;
+      margin-bottom: 0 !important;
 
-        .nav {
-            flex-wrap: nowrap;
-        }
+      &[aria-selected="false"] {
+        color: colors.$text-muted-color;
+        background-color: colors.$body-bg-var;
+      }
 
-        .nav-link {
-            color: colors.$text-emphasis-color;
-            font-size: sizes.$font-size-base;
-            text-align: center;
-            background-image: none !important;
-
-            &[aria-selected="false"] {
-                color: colors.$text-muted-color;
-                background-color: var(--bs-body-bg);
-                border-bottom: 1px solid var(--border-color-light);
-            }
-        }
-
-        .nav-item {
-            &:not(:last-child) .nav-link {
-                border-right: 1px solid var(--border-color-light);
-            }
-
-            &:first-child .nav-link {
-                border-top-left-radius: var(--bs-border-radius);
-            }
-
-            &:last-child .nav-link {
-                border-top-right-radius: var(--bs-border-radius);
-            }
-        }
+      &[aria-selected="true"] {
+        border-color: colors.$primary-var;
+        background-color: var(--faded-primary);
+      }
     }
+  }
+
+  // Segmented tab set (browser + client code) - mirrors .browser-tab-container
+  .segmented-tab-container {
+    border-radius: var(--bs-border-radius);
+    border: 1px solid var(--border-color-light);
+    background-color: var(--panel-bg-2);
+
+    .nav {
+      flex-wrap: nowrap;
+    }
+
+    .nav-link {
+      color: colors.$text-emphasis-color;
+      font-size: sizes.$font-size-base;
+      text-align: center;
+      background-image: none !important;
+
+      &[aria-selected="false"] {
+        color: colors.$text-muted-color;
+        background-color: var(--bs-body-bg);
+        border-bottom: 1px solid var(--border-color-light);
+      }
+    }
+
+    .nav-item {
+      &:not(:last-child) .nav-link {
+        border-right: 1px solid var(--border-color-light);
+      }
+
+      &:first-child .nav-link {
+        border-top-left-radius: var(--bs-border-radius);
+      }
+
+      &:last-child .nav-link {
+        border-top-right-radius: var(--bs-border-radius);
+      }
+    }
+  }
 }
